Add new Whois API client
